Refactor admin dashboard components: remove logout button, project form, and blog form; implement new blog and project management dialogs; enhance message handling in dashboard; update project and blog submission logic.
